Make health config independent from metrics
This allows the health/metric additions to be turned of without
influencing the other actuator feature.
