+++ grpc-client-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -2,6 +2,7 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 net.devh.boot.grpc.client.autoconfigure.GrpcClientAutoConfiguration,\
 net.devh.boot.grpc.client.autoconfigure.GrpcClientMetricAutoConfiguration,\
+net.devh.boot.grpc.client.autoconfigure.GrpcClientHealthAutoConfiguration,\
 net.devh.boot.grpc.client.autoconfigure.GrpcClientSecurityAutoConfiguration,\
 net.devh.boot.grpc.client.autoconfigure.GrpcClientTraceAutoConfiguration,\
 net.devh.boot.grpc.client.autoconfigure.GrpcDiscoveryClientAutoConfiguration
